requirements.txt
redis>=4.5.1
PyQt6>=6.4.0
pyinstaller>=5.7.0
pillow>=9.0.0 